LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/inline-view-timeline-current-time.tentative-expected.txt


FAIL View timeline with start and end scroll offsets that do not align with the scroll boundaries assert_equals: Effect is in the after phase at effect end time expected "1" but got "0.7"
FAIL View timeline does not clamp starting scroll offset at 0 assert_equals: Effect at the end of the active phase expected "1" but got "0.7"
PASS View timeline does not clamp end scroll offset at max scroll
FAIL View timeline with container having RTL layout assert_equals: Effect is in the after phase at effect end time expected "1" but got "0.7"
